addressing minor formatting and linter problems
